requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.16.3
MarkupSafe==3.0.2
multidict==6.3.2
packaging==24.2
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
